back/requirements.txt
Flask==1.0.2
Pyrebase==3.0.27
python-dotenv==0.10.2
